cleanup resolver and add some more tests
