+++ packages/sandpack-core/src/resolver/constants.ts
@@ -1,0 +1,1 @@
+export const EMPTY_SHIM = '//empty.js';
