+++ packages/sandpack-core/src/resolver/utils/pkg-json.test.ts
@@ -1,0 +1,33 @@
+import fs from 'fs';
+import path from 'path';
+
+import { processPackageJSON } from './pkg-json';
+
+const FIXTURE_PATH = path.join(__dirname, '../fixture');
+
+describe('process package.json', () => {
+  it('Should correctly process pkg.exports from @babel/runtime', () => {
+    const content = JSON.parse(
+      fs.readFileSync(
+        path.join(FIXTURE_PATH, 'node_modules/@babel/runtime/package.json'),
+        'utf-8'
+      )
+    );
+    const processedPkg = processPackageJSON(
+      content,
+      '/node_modules/@babel/runtime'
+    );
+    expect(processedPkg).toMatchSnapshot();
+  });
+
+  it('Should correctly handle nested pkg#exports fields (solid-js)', () => {
+    const content = JSON.parse(
+      fs.readFileSync(
+        path.join(FIXTURE_PATH, 'node_modules/solid-js/package.json'),
+        'utf-8'
+      )
+    );
+    const processedPkg = processPackageJSON(content, '/node_modules/solid-js');
+    expect(processedPkg).toMatchSnapshot();
+  });
+});
